- Implement placeholder for searchPage (Testing purposes)
